[tools] tools lib traceevent: Add API to read time information from kbuffer
Message-id: <20201106141119.29144-308-mpetlan@redhat.com>
Patchwork-id: 338404
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 307/329] tools lib traceevent: Add API to read time information from kbuffer
Bugzilla: 1873270
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873270
https://bugzilla.redhat.com/show_bug.cgi?id=1873270

upstream
========
commit 374855c5e4a76f7cc7f9aeb613c54929510eff18
Author: Steven Rostedt (Red Hat) <rostedt@goodmis.org>
Date: Thu Jul 2 14:53:45 2020 -0400

description
===========
Add the functions kbuffer_subbuf_timestamp() and kbuffer_ptr_delta() to
get the timing data stored in the ring buffer that is used to produced
the time stamps of the records.

This is useful for tools like trace-cmd to be able to display the
content of the read data to understand why the records show the time
stamps that they do.

[ Ported from trace-cmd.git ]
Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>