[tools] perf top/report: Fix infinite loop in the TUI for grouped events
Message-id: <20201106141119.29144-321-mpetlan@redhat.com>
Patchwork-id: 338416
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 320/329] perf top/report: Fix infinite loop in the TUI for grouped events
Bugzilla: 1873270
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873270
https://bugzilla.redhat.com/show_bug.cgi?id=1873270

upstream
========
commit d4ccbacb9c217fefb4332a9af81b785690cf1053
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Mon Aug 31 16:17:00 2020 -0300

description
===========
For a while we need to have a dummy event for doing things like
receiving PERF_RECORD_COMM, PERF_RECORD_EXEC, etc for threads being
created and dying while we synthesize the pre-existing ones at tool
start.

This 'dummy' event is needed for keeping track of thread lifetime events
early in the session but are uninteresting otherwise, i.e. no need to
have it in a initial events menu for the non-grouped case, i.e. for:

 # perf top -e cycles,instructions

or even for plain:

 # perf top

When 'cycles' and that 'dummy' event are in place.

The code to remove that 'dummy' event ended up creating an endless loop
for the grouped case, i.e.:

 # perf top -e '{cycles,instructions}'

Fix it.

Fixes: bee9ca1c8a237ca1 ("perf report TUI: Remove needless 'dummy' event from menu")
Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>