[powerpc] kvm: ppc: book3s hv: Set LPCR[HDICE] before writing HDEC
Message-id: <20201117152733.1509856-35-gkurz@redhat.com>
Patchwork-id: 339719
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 kernel virt PATCH 34/41] KVM: PPC: Book3S HV: Set LPCR[HDICE] before writing HDEC
Bugzilla: 1882796
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Paul Mackerras <paulus@ozlabs.org>

POWER8 and POWER9 machines have a hardware deviation where generation
of a hypervisor decrementer exception is suppressed if the HDICE bit
in the LPCR register is 0 at the time when the HDEC register
decrements from 0 to -1.  When entering a guest, KVM first writes the
HDEC register with the time until it wants the CPU to exit the guest,
and then writes the LPCR with the guest value, which includes
HDICE = 1.  If HDEC decrements from 0 to -1 during the interval
between those two events, it is possible that we can enter the guest
with HDEC already negative but no HDEC exception pending, meaning that
no HDEC interrupt will occur while the CPU is in the guest, or at
least not until HDEC wraps around.  Thus it is possible for the CPU to
keep executing in the guest for a long time; up to about 4 seconds on
POWER8, or about 4.46 years on POWER9 (except that the host kernel
hard lockup detector will fire first).

To fix this, we set the LPCR[HDICE] bit before writing HDEC on guest
entry.

Signed-off-by: Paul Mackerras <paulus@ozlabs.org>
(cherry picked from commit 35dfb43c243b28658ef757a2b9c6454f99cfc43c)
Signed-off-by: Greg Kurz <gkurz@redhat.com>

Conflicts:
	arch/powerpc/kvm/book3s_hv.c

This patch supsersedes downstream 4cc746ffc7b27 which was a temporary
RH-only fix (BZ: 1733467).

Signed-off-by: Jan Stancek <jstancek@redhat.com>