[kernel] kheaders: Do not regenerate archive if config is not changed
Message-id: <20201120165722.1735893-6-jolsa@redhat.com>
Patchwork-id: 340361
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.4 BZ1899585 05/15] kheaders: Do not regenerate archive if config is not changed
Bugzilla: 1899585
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>
RH-Acked-by: Toke Høiland-Jørgensen <toke@redhat.com>

Bugzilla: 1899585
https://bugzilla.redhat.com/show_bug.cgi?id=1899585

upstream
========
commit 1457dc9ed8da871fbbc0a2ebdaed0405eeeed0cf
Author: Joel Fernandes (Google) <joel@joelfernandes.org>
Date: Wed May 15 17:35:52 2019 -0400

description
===========
Linus reported an issue that doing an allmodconfig was causing the
kheaders archive to be regenerated even though the config is the same.
This patch fixes the issue by ignoring the config-related header files
for "knowing when to regenerate based on timestamps".  Instead, if the
CONFIG_X_Y option really changes, then we there are the
include/config/X/Y.h which will already tells us "if a config really
changed". So we don't really need these files for regeneration detection
anyway, and ignoring them fixes Linus's issue.

Signed-off-by: Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>