[kernel] kheaders: substituting --sort in archive creation
Message-id: <20201120165722.1735893-10-jolsa@redhat.com>
Patchwork-id: 340371
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.4 BZ1899585 09/15] kheaders: substituting --sort in archive creation
Bugzilla: 1899585
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>
RH-Acked-by: Toke Høiland-Jørgensen <toke@redhat.com>

Bugzilla: 1899585
https://bugzilla.redhat.com/show_bug.cgi?id=1899585

upstream
========
commit 700dea5a0bea9f64eba89fae7cb2540326fdfdc1
Author: Dmitry Goldin <dgoldin+lkml@protonmail.ch>
Date: Wed Oct 9 13:42:14 2019 +0000

description
===========
The option --sort=ORDER was only introduced in tar 1.28 (2014), which
is rather new and might not be available in some setups.

This patch tries to replicate the previous behaviour as closely as
possible to fix the kheaders build for older environments. It does
not produce identical archives compared to the previous version due
to minor sorting differences but produces reproducible results itself
in my tests.

Signed-off-by: Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>