[kernel] kheaders: optimize md5sum calculation for in-tree builds
Message-id: <20201120165722.1735893-12-jolsa@redhat.com>
Patchwork-id: 340366
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.4 BZ1899585 11/15] kheaders: optimize md5sum calculation for in-tree builds
Bugzilla: 1899585
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>
RH-Acked-by: Toke Høiland-Jørgensen <toke@redhat.com>

Bugzilla: 1899585
https://bugzilla.redhat.com/show_bug.cgi?id=1899585

upstream
========
commit 0e11773e76098729552b750ccff79374d1e62002
Author: Masahiro Yamada <yamada.masahiro@socionext.com>
Date: Tue Oct 8 21:05:53 2019 +0900

description
===========
This script computes md5sum of headers in srctree and in objtree.
However, when we are building in-tree, we know the srctree and the
objtree are the same. That is, we end up with the same computation
twice. In fact, the first two lines of kernel/kheaders.md5 are always
the same for in-tree builds.

Unify the two md5sum calculations.

For in-tree builds ($building_out_of_srctree is empty), we check
only two directories, "include", and "arch/$SRCARCH/include".

For out-of-tree builds ($building_out_of_srctree is 1), we check
4 directories, "$srctree/include", "$srctree/arch/$SRCARCH/include",
"include", and "arch/$SRCARCH/include" since we know they are all
different.

Signed-off-by: Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>