[kernel] kheaders: optimize header copy for in-tree builds
Message-id: <20201120165722.1735893-13-jolsa@redhat.com>
Patchwork-id: 340367
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.4 BZ1899585 12/15] kheaders: optimize header copy for in-tree builds
Bugzilla: 1899585
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>
RH-Acked-by: Toke Høiland-Jørgensen <toke@redhat.com>

Bugzilla: 1899585
https://bugzilla.redhat.com/show_bug.cgi?id=1899585

upstream
========
commit ea79e5168be644fdaf7d4e6a73eceaf07b3da76a
Author: Masahiro Yamada <yamada.masahiro@socionext.com>
Date: Tue Oct 8 21:05:54 2019 +0900

description
===========
This script copies headers by the cpio command twice; first from
srctree, and then from objtree. However, when we building in-tree,
we know the srctree and the objtree are the same. That is, all the
headers copied by the first cpio are overwritten by the second one.

Skip the first cpio when we are building in-tree.

Signed-off-by: Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>