[kernel] kheaders: remove the last bashism to allow sh to run it
Message-id: <20201120165722.1735893-14-jolsa@redhat.com>
Patchwork-id: 340368
Patchwork-instance: patchwork
O-Subject: [PATCH RHEL8.4 BZ1899585 13/15] kheaders: remove the last bashism to allow sh to run it
Bugzilla: 1899585
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>
RH-Acked-by: Toke Høiland-Jørgensen <toke@redhat.com>

Bugzilla: 1899585
https://bugzilla.redhat.com/show_bug.cgi?id=1899585

upstream
========
commit 1463f74f492eea7191f0178e01f3d38371a48210
Author: Masahiro Yamada <yamada.masahiro@socionext.com>
Date: Tue Oct 8 21:05:55 2019 +0900

RHEL8
=====
We did not backport:
  858805b336be kbuild: add $(BASH) to run scripts with bash-extension

because this patch reverts it back, hence the indentation
conflict in one of the hunks.

description
===========
'pushd' ... 'popd' is the last bash-specific code in this script.
One way to avoid it is to run the code in a sub-shell.

With that addressed, you can run this script with sh.

I replaced $(BASH) with $(CONFIG_SHELL), and I changed the hashbang
to #!/bin/sh.

Signed-off-by: Jiri Olsa <jolsa@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>