[arm64] kvm: Annotate guest entry/exit as a single function
Message-id: <20201125134123.18590-3-eric.auger@redhat.com>
Patchwork-id: 341715
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 virt BZ1882794 PATCH v2 02/91] arm64: KVM: Annotate guest entry/exit as a single function
Bugzilla: 1882794
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Gavin Shan <gshan@redhat.com>

From: Mark Brown <broonie@kernel.org>

In an effort to clarify and simplify the annotations of assembly
functions in the kernel new macros have been introduced replacing ENTRY
and ENDPROC. There are separate annotations SYM_FUNC_ for normal C
functions and SYM_CODE_ for other code. Currently __guest_enter and
__guest_exit are annotated as standard functions but this is not
entirely correct as the former doesn't do a normal return and the latter
is not entered in a normal fashion. From the point of view of the
hypervisor the guest entry/exit may be viewed as a single
function which happens to have an eret in the middle of it so let's
annotate it as such.

Suggested-by: Mark Rutland <mark.rutland@arm.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Signed-off-by: Marc Zyngier <maz@kernel.org>
Acked-by: Will Deacon <will@kernel.org>
Link: https://lore.kernel.org/r/20200120124706.8681-1-broonie@kernel.org
(cherry picked from commit 6645d8542ef922486b733d415d2bec3b0622c27e)
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>