[arm64] kvm: arm64: Add kvm_extable for vaxorcism code
Message-id: <20201125134123.18590-90-eric.auger@redhat.com>
Patchwork-id: 341802
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 virt BZ1882794 PATCH v2 89/91] KVM: arm64: Add kvm_extable for vaxorcism code
Bugzilla: 1882794
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Gavin Shan <gshan@redhat.com>

From: James Morse <james.morse@arm.com>

KVM has a one instruction window where it will allow an SError exception
to be consumed by the hypervisor without treating it as a hypervisor bug.
This is used to consume asynchronous external abort that were caused by
the guest.

As we are about to add another location that survives unexpected exceptions,
generalise this code to make it behave like the host's extable.

KVM's version has to be mapped to EL2 to be accessible on nVHE systems.

The SError vaxorcism code is a one instruction window, so has two entries
in the extable. Because the KVM code is copied for VHE and nVHE, we end up
with four entries, half of which correspond with code that isn't mapped.

Signed-off-by: James Morse <james.morse@arm.com>
Reviewed-by: Marc Zyngier <maz@kernel.org>
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
(cherry picked from commit e9ee186bb735bfc17fa81dbc9aebf268aee5b41e)
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>