[fs] cifs: misc: Use array_size() in if-statement controlling expression
Message-id: <20201203054653.24822-69-lsahlber@redhat.com>
Patchwork-id: 342834
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 FS BZ 1903883 68/85] cifs: misc: Use array_size() in if-statement controlling expression
Bugzilla: 1903883
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: "Gustavo A. R. Silva" <gustavoars@kernel.org>

Bugzilla: 1903883
Build: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=33455430

Use array_size() instead of the open-coded version in the controlling
expression of the if statement.

Also, while there, use the preferred form for passing a size of a struct.
The alternative form where struct name is spelled out hurts readability
and introduces an opportunity for a bug when the pointer variable type is
changed but the corresponding sizeof that is passed as argument is not.

This issue was found with the help of Coccinelle and, audited and fixed
manually.

Addresses-KSPP-ID: https://github.com/KSPP/linux/issues/83
Signed-off-by: Gustavo A. R. Silva <gustavoars@kernel.org>
Signed-off-by: Steve French <stfrench@microsoft.com>
Reviewed-by: Aurelien Aptel <aaptel@suse.com>
Reviewed-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Steve French <stfrench@microsoft.com>
(cherry picked from commit bf1028a41eaf0ce39518cbdda34cdb717f16364a)
Signed-off-by: Ronnie Sahlberg <lsahlber@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>