[x86] boot/compressed/64: Switch to __KERNEL_CS after GDT is loaded
Message-id: <20201204163312.2525587-5-vkuznets@redhat.com>
Patchwork-id: 343341
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 virt PATCH 004/102] x86/boot/compressed/64: Switch to __KERNEL_CS after GDT is loaded
Bugzilla: 1868080
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Joerg Roedel <jroedel@suse.de>

When the pre-decompression code loads its first GDT in startup_64(), it
is still running on the CS value of the previous GDT. In the case of
SEV-ES, this is the EFI GDT but it can be anything depending on what has
loaded the kernel (boot loader, container runtime, etc.)

To make exception handling work (especially IRET) the CPU needs to
switch to a CS value in the current GDT, so jump to __KERNEL_CS after
the first GDT is loaded. This is prudent also as a general sanitization
of CS to a known good value.

 [ bp: Massage commit message. ]

Signed-off-by: Joerg Roedel <jroedel@suse.de>
Signed-off-by: Borislav Petkov <bp@suse.de>
Link: https://lkml.kernel.org/r/20200428151725.31091-13-joro@8bytes.org
(cherry picked from commit 34bb49229f19399a5b45c323afb5749f31f7876c)

Conflicts:
	arch/x86/boot/compressed/head_64.S (context)
Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>