[x86] perf/x86/intel/lbr: Factor out a new struct for generic optimization
Message-id: <20201214095352.11924-20-mpetlan@redhat.com>
Patchwork-id: 344241
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 019/248] perf/x86/intel/lbr: Factor out a new struct for generic optimization
Bugzilla: 1811839 1731571 1838555 1838896 1873271
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873271
https://bugzilla.redhat.com/show_bug.cgi?id=1873271

upstream
========
commit 530bfff6480307d210734222a54d56af7f908957
Author: Kan Liang <kan.liang@linux.intel.com>
Date: Fri Jul 3 05:49:11 2020 -0700

description
===========
To reduce the overhead of a context switch with LBR enabled, some
generic optimizations were introduced, e.g. avoiding restore LBR if no
one else touched them. The generic optimizations can also be used by
Architecture LBR later. Currently, the fields for the generic
optimizations are part of structure x86_perf_task_context, which will be
deprecated by Architecture LBR. A new structure should be introduced
for the common fields of generic optimization, which can be shared
between Architecture LBR and model-specific LBR.

Both 'valid_lbrs' and 'tos' are also used by the generic optimizations,
but they are not moved into the new structure, because Architecture LBR
is stack-like. The 'valid_lbrs' which records the index of the valid LBR
is not required anymore. The TOS MSR will be removed.

LBR registers may be cleared in the deep Cstate. If so, the generic
optimizations should not be applied. Perf has to unconditionally
restore the LBR registers. A generic function is required to detect the
reset due to the deep Cstate. lbr_is_reset_in_cstate() is introduced.
Currently, for the model-specific LBR, the TOS MSR is used to detect the
reset. There will be another method introduced for Architecture LBR
later.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>