[tools] perf parse: Provide a way to pass a fake_pmu to parse_events()
Message-id: <20201214095352.11924-47-mpetlan@redhat.com>
Patchwork-id: 344276
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 046/248] perf parse: Provide a way to pass a fake_pmu to parse_events()
Bugzilla: 1811839 1731571 1838555 1838896 1873271
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873271
https://bugzilla.redhat.com/show_bug.cgi?id=1873271

upstream
========
commit 3bf91aa5aa493b2ac24ffcfc93ffe861bf03685f
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Jun 3 12:32:55 2020 -0300

description
===========
This is an alternative patch to what Jiri sent that instead of changing
all callers to parse_events() for allowing to pass a fake_pmu, provide
another function specifically for that.

>From Jiri's patch:

This way it's possible to parse events from PMUs which are not present
in the system. It's available only for testing purposes coming in
following changes, so all the current users set fake_pmu argument as
false.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>