[tools] perf cs-etm: Allow no CoreSight sink to be specified on command line
Message-id: <20201214095352.11924-62-mpetlan@redhat.com>
Patchwork-id: 344283
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 061/248] perf cs-etm: Allow no CoreSight sink to be specified on command line
Bugzilla: 1811839 1731571 1838555 1838896 1873271
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873271
https://bugzilla.redhat.com/show_bug.cgi?id=1873271

upstream
========
commit 47446212832871e66f3b5fbf021c148eced2e24c
Author: Mike Leach <mike.leach@linaro.org>
Date: Tue Jun 16 17:40:41 2020 +0100

description
===========
Adjust the handling of the session sink selection to allow no sink to be
selected on the command line. This then forwards the sink selection to
the CoreSight infrastructure which will attempt to select a sink based
on the default sink select priorities.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>