[tools] tools lib traceevent: Add builtin handler for trace_marker_raw
Message-id: <20201214095352.11924-89-mpetlan@redhat.com>
Patchwork-id: 344309
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 088/248] tools lib traceevent: Add builtin handler for trace_marker_raw
Bugzilla: 1811839 1731571 1838555 1838896 1873271
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873271
https://bugzilla.redhat.com/show_bug.cgi?id=1873271

upstream
========
commit 9b8179b38ba29c0c8fbfa09b2fe685edca0f6061
Author: Steven Rostedt (VMware) <rostedt@goodmis.org>
Date: Thu Jul 2 14:53:57 2020 -0400

description
===========
When something is written into trace_marker_raw, it goes in as a binary.
But the printk_fmt() of the event that is created (raw_data)'s format
file only prints the first byte of data:

  print fmt: "id:%04x %08x", REC->id, (int)REC->buf[0]

This is not very useful if we want to see the full data output.

Implement the processing of the raw_data event like it is in the kernel.

[ Ported from trace-cmd.git ]

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>