[tools] perf script: Show text poke address symbol
Message-id: <20201214095352.11924-98-mpetlan@redhat.com>
Patchwork-id: 344318
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 097/248] perf script: Show text poke address symbol
Bugzilla: 1811839 1731571 1838555 1838896 1873271
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873271
https://bugzilla.redhat.com/show_bug.cgi?id=1873271

upstream
========
commit 7eeb9855c1a4e6037bb1237a2a68ff1ee7aa4e67
Author: Adrian Hunter <adrian.hunter@intel.com>
Date: Tue May 12 15:19:22 2020 +0300

description
===========
It is generally more useful to show the symbol with an address. In this
case, the print function requires the 'machine' which means changing
callers to provide it as a parameter. It is optional because most events
do not need it and the callers that matter can provide it.

Committer notes:

Made 'union perf_event' continue to be the first parameter to the
perf_event__fprintf() and perf_event__fprintf_text_poke() events.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>