[tools] perf probe: Fix memory leakage when the probe point is not found
Message-id: <20201214095352.11924-103-mpetlan@redhat.com>
Patchwork-id: 344327
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 102/248] perf probe: Fix memory leakage when the probe point is not found
Bugzilla: 1811839 1731571 1838555 1838896 1873271
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873271
https://bugzilla.redhat.com/show_bug.cgi?id=1873271

upstream
========
commit 12d572e785b15bc764e956caaa8a4c846fd15694
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date: Fri Jul 10 22:11:23 2020 +0900

description
===========
Fix the memory leakage in debuginfo__find_trace_events() when the probe
point is not found in the debuginfo. If there is no probe point found in
the debuginfo, debuginfo__find_probes() will NOT return -ENOENT, but 0.

Thus the caller of debuginfo__find_probes() must check the tf.ntevs and
release the allocated memory for the array of struct probe_trace_event.

The current code releases the memory only if the debuginfo__find_probes()
hits an error but not checks tf.ntevs. In the result, the memory allocated
on *tevs are not released if tf.ntevs == 0.

This fixes the memory leakage by checking tf.ntevs == 0 in addition to
ret < 0.

Fixes: ff741783506c ("perf probe: Introduce debuginfo to encapsulate dwarf information")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>