[tools] libtraceevent: Fixed description of tep_add_plugin_path() API
Message-id: <20201214095352.11924-167-mpetlan@redhat.com>
Patchwork-id: 344378
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 166/248] libtraceevent: Fixed description of tep_add_plugin_path() API
Bugzilla: 1811839 1731571 1838555 1838896 1873271
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873271
https://bugzilla.redhat.com/show_bug.cgi?id=1873271

upstream
========
commit db885ed481b33bb4b782ca9a76841bc9ec0e2d8f
Author: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Date: Tue Jul 21 21:16:50 2020 -0400

description
===========
Changed the description of tep_add_plugin_path() API to reflect the
logic of the function. The suffix of plugin files is not hardcoded
to ".so", it depends on the custom plugin loader callback.

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>