[tools] perf metric: Do not free metric when failed to resolve
Message-id: <20201214095352.11924-244-mpetlan@redhat.com>
Patchwork-id: 344470
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH 243/248] perf metric: Do not free metric when failed to resolve
Bugzilla: 1811839 1731571 1838555 1838896 1873271
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1873271
https://bugzilla.redhat.com/show_bug.cgi?id=1873271

upstream
========
commit 6f47ed6cd12a0ea9b55b19b5d6e4a3b490a97bdf
Author: Namhyung Kim <namhyung@kernel.org>
Date: Tue Sep 15 12:18:17 2020 +0900

description
===========
It's dangerous to free the original metric when it's called from
resolve_metric() as it's already in the metric_list and might have other
resources too.  Instead, it'd better let them bail out and be released
properly at the later stage.

So add a check when it's called from metricgroup__add_metric() and
release it.  Also make sure that mp is set properly.

Fixes: 83de0b7d535de ("perf metric: Collect referenced metrics in struct metric_ref_node")

Signed-off-by: Michael Petlan <mpetlan@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>