[powerpc] powerpc/pci: Remove LSI mappings on device teardown
Message-id: <20210105182436.22803-3-didoming@redhat.com>
Patchwork-id: 347092
Patchwork-instance: patchwork
O-Subject: [RHEL8.4 PATCH BZ1909722 0002/0002] powerpc/pci: Remove LSI mappings on device teardown
Bugzilla: 1909722
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Bugzilla: http://bugzilla.redhat.com/1909722
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=34041527

When a passthrough IO adapter is removed from a pseries machine using hash
MMU and the XIVE interrupt mode, the POWER hypervisor expects the guest OS
to clear all page table entries related to the adapter. If some are still
present, the RTAS call which isolates the PCI slot returns error 9001
"valid outstanding translations" and the removal of the IO adapter fails.
This is because when the PHBs are scanned, Linux maps automatically the
INTx interrupts in the Linux interrupt number space but these are never
removed.

This problem can be fixed by adding the corresponding unmap operation when
the device is removed. There's no pcibios_* hook for the remove case, but
the same effect can be achieved using a bus notifier.

Because INTx are shared among PHBs (and potentially across the system),
this adds tracking of virq to unmap them only when the last user is gone.

[aik: added refcounter]

commit 450be4960a0fb89b931a6bb3c3f0bb538ac4c03c

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Tested-by: Cédric Le Goater <clg@kaod.org>
Reviewed-by: Frederic Barrat <fbarrat@linux.ibm.com>
Reviewed-by: Cédric Le Goater <clg@kaod.org>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/20201202005222.5477-1-aik@ozlabs.ru
Signed-off-by: Diego Domingos <didoming@redhat.com>
Signed-off-by: Jan Stancek <jstancek@redhat.com>